fix: fetch 문제, controlled 문제 해결

+++ frontend/src/context/ProductsContext.tsx
@@ -62,6 +62,9 @@ export const useProductItem = (id: number) => {
   const [product, setProduct] = useState<Product | null>(null);
 
   useEffect(() => {
+    if(!id || id === 0) {
+      return;
+    }
     apiFetch(`/api/v1/products/${id}`).then(setProduct)
     .catch((error) => {
       alert(`${error.resultCode} : ${error.msg}`);
@@ -69,6 +72,9 @@ export const useProductItem = (id: number) => {
   }, []);
 
   const deleteProduct = (onSuccess: (data: any) => void) => {
+    if(!id || id === 0) {
+      return;
+    }
     apiFetch(`/api/v1/adm/products/${id}`, {
       method: "DELETE"
     }).then(onSuccess)
@@ -85,6 +91,9 @@ export const useProductItem = (id: number) => {
     imageUrl: string,
     onSuccess:(data: any) => void
   }) => {
+    if(!id || id === 0) {
+      return;
+    }
     apiFetch(`/api/v1/adm/products/${id}`, {
       method: "PUT",
       body: JSON.stringify({
@@ -103,13 +112,19 @@ export const useProductImage = (productId: number) => {
   const [productImages, setProductImages] = useState<ProductImage[] | null>(null);
 
   useEffect(() => {
+    if(!productId || productId === 0) {
+      return;
+    }
     apiFetch(`/api/v1/products/${productId}/images`).then(setProductImages)
     .catch((error) => {
       alert(`${error.resultCode} : ${error.msg}`);
     });
-  }, []);
+  }, [productId]);
 
-  const addProductImage = (url) => {
+  const addProductImage = (url: string, productId: number) => {
+    if(!productId || productId === 0) {
+      return;
+    }
     apiFetch(`/api/v1/adm/products/${productId}/images`, {
       method: "POST",
       body: JSON.stringify({url}),
@@ -128,6 +143,9 @@ export const useProductImageItem = (productId: number, id: number) => {
   const [productImage, setProductImage] = useState<ProductImage>();
 
   useEffect(() => {
+    if(!productId || productId === 0) {
+      return;
+    }
     apiFetch(`/api/v1/products/${productId}/images/${id}`).then(setProductImage)
     .catch((error) => {
       alert(`${error.resultCode} : ${error.msg}`);
@@ -135,6 +153,9 @@ export const useProductImageItem = (productId: number, id: number) => {
   }, []);
 
   const deleteProductImage = (onSuccess: (data: any) => void) => {
+    if(!productId || productId === 0) {
+      return;
+    }
     apiFetch(`/api/v1/adm/products/${productId}/images/${id}`, {
       method: "DELETE"
     }).then(onSuccess)
@@ -147,6 +168,9 @@ export const useProductImageItem = (productId: number, id: number) => {
     url: string,
     onSuccess:(data: any) => void
   }) => {
+    if(!productId || productId === 0) {
+      return;
+    }
     apiFetch(`/api/v1/adm/products/${productId}/images/${id}`, {
       method: "PUT",
       body: JSON.stringify({url}),
